Updated readme.md for regular webpage version
Added a new gif to show the search function.

I commented out Theme information as it looks like this feature is not available now however can be included in the readme in the future if this feature is added.

Unsure about Sync item in TODO category. It may not be relevant but I left it for now because I don't know what it is.